Fix some bugs

+++ controllers/contactsControllers.js
@@ -1,20 +1,18 @@
 import contactsService from "../services/contactsServices.js";
-
 import HttpError from "../helpers/HttpError.js";
 import {
   createContactSchema,
   updateContactSchema,
   favoriteContactSchema,
 } from "../schemas/contactsSchemas.js";
+import mongoose from "mongoose";
 
 export const getAllContacts = async (req, res, next) => {
   try {
     const result = await contactsService.getAllContacts();
-
     if (!result) {
-      throw new HttpError(404, "Not found");
+      throw new HttpError(404, "Contacts not found");
     }
-
     res.json(result);
   } catch (error) {
     next(error);
@@ -24,6 +22,11 @@ export const getAllContacts = async (req, res, next) => {
 export const getOneContact = async (req, res, next) => {
   try {
     const { id } = req.params;
+
+    if (!mongoose.isValidObjectId(id)) {
+      throw new HttpError(404, "Not found");
+    }
+
     const result = await contactsService.getOneContact(id);
 
     if (!result) {
@@ -38,9 +41,14 @@ export const getOneContact = async (req, res, next) => {
 
 export const deleteContact = async (req, res, next) => {
   try {
+    const { id } = req.params;
+
+    if (!mongoose.isValidObjectId(id)) {
+      throw new HttpError(404, "Not found");
+    }
     const result = await contactsService.deleteContact(req.params.id);
     if (!result) {
-      throw new HttpError(404, "Not found");
+      throw new HttpError(404, "Contact not found");
     }
     res.json(result);
   } catch (error) {
@@ -55,7 +63,6 @@ export const createContact = async (req, res, next) => {
       throw new HttpError(400, error.message);
     }
     const result = await contactsService.createContact(req.body);
-
     res.status(201).json(result);
   } catch (error) {
     next(error);
@@ -71,7 +78,7 @@ export const updateContact = async (req, res, next) => {
     const { id } = req.params;
     const result = await contactsService.updateContact(id, req.body);
     if (!result) {
-      throw new HttpError(404, "Not found");
+      throw new HttpError(404, "Contact not found");
     }
     res.json(result);
   } catch (error) {
@@ -83,7 +90,6 @@ export const updateStatusContact = async (req, res, next) => {
   try {
     const { id } = req.params;
     const { favorite } = req.body;
-
     const { error } = favoriteContactSchema.validate({ favorite });
     if (error) {
       throw new HttpError(400, error.message);
@@ -91,13 +97,10 @@ export const updateStatusContact = async (req, res, next) => {
     if (favorite === undefined || typeof favorite !== "boolean") {
       throw new HttpError(400, "Invalid favorite value");
     }
-
     const result = await contactsService.updateStatusContact(id, favorite);
-
     if (!result) {
-      throw new HttpError(404, "Not found");
+      throw new HttpError(404, "Contact not found");
     }
-
     res.json(result);
   } catch (error) {
     next(error);
